requirements.txt
flask==1.1.1
sqlalchemy==1.3.15
flask-sqlalchemy==2.4.1
psycopg2-binary==2.8.4
sqlalchemy-utils==0.36.3
flask-wtf==0.14.3
requests==2.23.0
redis
celery==4.4.2
gunicorn
pynsq
marshmallow==3.5.1
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.22.3
Flask-RESTful
Flask-Cors==3.0.8
flask-marshmallow==0.11.0
